Refactoring search tasks - moving search task related internal class and methods to partial class - extracting methods
