+++ ConfluenceEX/ViewModel/BeforeSignInViewModel.cs
@@ -1,15 +1,11 @@
 ﻿using ConfluenceEX.Command;
 using ConfluenceEX.Common;
+using ConfluenceEX.Main;
 using ConfluenceRESTClient.Service;
 using ConfluenceRESTClient.Service.Implementation;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Runtime.InteropServices;
 using System.Security;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Input;
 
 namespace ConfluenceEX.ViewModel
 {
